GPS_RECIEVER FINAL TOUCHES

+++ src/Math_Functions.c
@@ -2,14 +2,14 @@
 const double Earth_Radius = 6371000;
 const float PI = 3.14159;
 
-float Todegree(float angle)
+double Todegree(double angle)
 {
 	int degree = (int)angle / 100;
-	float minutes = angle - (float)degree * 100;
+	double minutes = angle - (double)degree * 100;
 	return (degree + (minutes / 60));
 }
 
-float Torad(float angle)
+double Torad(double angle)
 {
 	return (angle * PI / 180);
 }
